feat: finish model
