+++ app/src/main/java/dog/snow/androidrecruittest/data/model/address/Street.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.address
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Street(val value: String) : Parcelable
