+++ app/src/main/java/dog/snow/androidrecruittest/data/model/address/Suite.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.address
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Suite(val value: String) : Parcelable
