+++ app/src/main/java/dog/snow/androidrecruittest/data/model/address/ZipCode.kt
@@ -1,4 +1,4 @@
-package dog.snow.androidrecruittest.data.model.type
+package dog.snow.androidrecruittest.data.model.address
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
