+++ app/src/main/java/dog/snow/androidrecruittest/data/model/album/RawAlbum.kt
@@ -2,7 +2,7 @@ package dog.snow.androidrecruittest.data.model.album
 
 import android.os.Parcelable
 import dog.snow.androidrecruittest.data.model.common.Id
-import dog.snow.androidrecruittest.data.model.album.Title
+import dog.snow.androidrecruittest.data.model.photo.Title
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
