+++ app/src/main/java/dog/snow/androidrecruittest/data/model/company/Bs.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.company
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Bs(val value: String) : Parcelable
