+++ app/src/main/java/dog/snow/androidrecruittest/data/model/geolocation/RawGeo.kt
@@ -1,0 +1,7 @@
+package dog.snow.androidrecruittest.data.model.geolocation
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class RawGeo(val lat: Coordinate, val lng: Coordinate) : Parcelable
