+++ app/src/main/java/dog/snow/androidrecruittest/data/model/photo/RawPhoto.kt
@@ -2,8 +2,6 @@ package dog.snow.androidrecruittest.data.model.photo
 
 import android.os.Parcelable
 import dog.snow.androidrecruittest.data.model.common.Id
-import dog.snow.androidrecruittest.data.model.album.Title
-import dog.snow.androidrecruittest.data.model.photo.Url
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
