+++ app/src/main/java/dog/snow/androidrecruittest/data/model/user/RawUser.kt
@@ -1,10 +1,11 @@
 package dog.snow.androidrecruittest.data.model.user
 
 import android.os.Parcelable
+import dog.snow.androidrecruittest.data.model.address.RawAddress
 import dog.snow.androidrecruittest.data.model.common.Id
 import dog.snow.androidrecruittest.data.model.common.Name
-import dog.snow.androidrecruittest.data.model.geolocation.Coordinate
-import dog.snow.androidrecruittest.data.model.type.*
+import dog.snow.androidrecruittest.data.model.company.RawCompany
+
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
@@ -17,23 +18,4 @@ data class RawUser(
     val phone: Phone,
     val website: Website,
     val company: RawCompany
-) : Parcelable {
-    @Parcelize
-    data class RawAddress(
-        val street: String,
-        val suite: String,
-        val city: City,
-        val zipcode: ZipCode,
-        val geo: RawGeo
-    ) : Parcelable {
-        @Parcelize
-        data class RawGeo(val lat: Coordinate, val lng: Coordinate) : Parcelable
-    }
-
-    @Parcelize
-    data class RawCompany(
-        val name: Name,
-        val catchPhrase: String,
-        val bs: String
-    ) : Parcelable
-}
+) : Parcelable
